src/test/javascript/spec/app/entities/base/market/market-update.component.spec.ts
/* tslint:disable max-line-length */
import { shallowMount, createLocalVue, Wrapper } from '@vue/test-utils';
import sinon, { SinonStubbedInstance } from 'sinon';
import Router from 'vue-router';
import { ToastPlugin } from 'bootstrap-vue';

import * as config from '@/shared/config/config';
import MarketUpdateComponent from '@/entities/base/market/market-update.vue';
import MarketClass from '@/entities/base/market/market-update.component';
import MarketService from '@/entities/base/market/market.service';

import AlertService from '@/shared/alert/alert.service';

const localVue = createLocalVue();

config.initVueApp(localVue);
const store = config.initVueXStore(localVue);
const router = new Router();
localVue.use(Router);
localVue.use(ToastPlugin);
localVue.component('font-awesome-icon', {});
localVue.component('b-input-group', {});
localVue.component('b-input-group-prepend', {});
localVue.component('b-form-datepicker', {});
localVue.component('b-form-input', {});

describe('Component Tests', () => {
  describe('Market Management Update Component', () => {
    let wrapper: Wrapper<MarketClass>;
    let comp: MarketClass;
    let marketServiceStub: SinonStubbedInstance<MarketService>;

    beforeEach(() => {
      marketServiceStub = sinon.createStubInstance<MarketService>(MarketService);

      wrapper = shallowMount<MarketClass>(MarketUpdateComponent, {
        store,
        localVue,
        router,
        provide: {
          marketService: () => marketServiceStub,
          alertService: () => new AlertService(),
        },
      });
      comp = wrapper.vm;
    });

    describe('save', () => {
      it('Should call update service on save for existing entity', async () => {
        // GIVEN
        const entity = { id: 123 };
        comp.market = entity;
        marketServiceStub.update.resolves(entity);

        // WHEN
        comp.save();
        await comp.$nextTick();

        // THEN
        expect(marketServiceStub.update.calledWith(entity)).toBeTruthy();
        expect(comp.isSaving).toEqual(false);
      });

      it('Should call create service on save for new entity', async () => {
        // GIVEN
        const entity = {};
        comp.market = entity;
        marketServiceStub.create.resolves(entity);

        // WHEN
        comp.save();
        await comp.$nextTick();

        // THEN
        expect(marketServiceStub.create.calledWith(entity)).toBeTruthy();
        expect(comp.isSaving).toEqual(false);
      });
    });

    describe('Before route enter', () => {
      it('Should retrieve data', async () => {
        // GIVEN
        const foundMarket = { id: 123 };
        marketServiceStub.find.resolves(foundMarket);
        marketServiceStub.retrieve.resolves([foundMarket]);

        // WHEN
        comp.beforeRouteEnter({ params: { marketId: 123 } }, null, cb => cb(comp));
        await comp.$nextTick();

        // THEN
        expect(comp.market).toBe(foundMarket);
      });
    });

    describe('Previous state', () => {
      it('Should go previous state', async () => {
        comp.previousState();
        await comp.$nextTick();

        expect(comp.$router.currentRoute.fullPath).toContain('/');
      });
    });
  });
});
